weather_data/requirements.txt
asgiref==3.5.2
Django==3.2.16
djangorestframework==3.14.0
greenlet==2.0.1
hurry==1.1
importlib-metadata==5.1.0
mysqlclient==2.0.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.11.0